#147 | feat(tests): separate integer and long values tests
