+++ src/test/groovy/pl/allegro/finance/tradukisto/internal/languages/english/EnglishIntegerValuesTest.groovy
@@ -1,22 +1,18 @@
 package pl.allegro.finance.tradukisto.internal.languages.english
 
-import pl.allegro.finance.tradukisto.internal.languages.AbstractValuesTest
+import pl.allegro.finance.tradukisto.internal.languages.AbstractIntegerValuesTest
 
 import static pl.allegro.finance.tradukisto.internal.Container.englishContainer
 
-class EnglishValuesTest extends AbstractValuesTest {
+class EnglishIntegerValuesTest extends AbstractIntegerValuesTest {
 
     def setup() {
         intConverter = englishContainer().integerConverter
-        longConverter = englishContainer().longConverter
     }
 
     @Override
-    ValuesTestData getTestData() {
-        testData = new ValuesTestData(
-                integerTranslations,
-                longTranslations
-        )
+    IntegerValuesTestData getTestData() {
+        testData = new IntegerValuesTestData(integerTranslations)
     }
 
     private static integerTranslations = [
@@ -106,24 +102,5 @@ class EnglishValuesTest extends AbstractValuesTest {
             1_000_000_000: "one billion",
             2_147_483_647: "two billion one hundred forty-seven million four hundred eighty-three thousand six hundred " +
                     "forty-seven"
-    ]
-
-    private static longTranslations = [
-            5_000_000_000            : "five billion",
-
-            1_000_000_000_000        : "one trillion",
-            2_000_000_000_000        : "two trillion",
-            5_000_000_000_000        : "five trillion",
-
-            1_000_000_000_000_000    : "one quadrillion",
-            2_000_000_000_000_000    : "two quadrillion",
-            5_000_000_000_000_000    : "five quadrillion",
-
-            1_000_000_000_000_000_000: "one quintillion",
-            2_000_000_000_000_000_000: "two quintillion",
-            (Long.MAX_VALUE)         : "nine quintillion two hundred twenty-three quadrillion " +
-                    "three hundred seventy-two trillion thirty-six billion " +
-                    "eight hundred fifty-four million seven hundred seventy-five thousand " +
-                    "eight hundred seven"
     ]
 }
